WrappedReader only wraps methods which are valid after initFile/setId
This makes it easier to pass the iunchanged parent behaviour to child classes.
In order to avoid changes to FilePatternReader these methods are therefore overriden.
nopte this also restore the pre-existing behaviour of FilePatternReader in ignoring IMetadataConfigurable methods.
